refactor(desktop): invert the shared/features imports the planner move left behind
The settings/targets decoupling reduced the feature cycle but did not break it,
and it introduced four `shared/ -> features/` imports -- the direction the move
exists to remove.

- Delete `shared/settings-queries.ts`. `settingsQueryOptions` had no call site,
  no test, and was the only reason `shared/` imported `features/settings`.
- Move `planner-catalog.ts` (pure data over `@/bindings` + `@/lib/i18n`) to
  `shared/planner/`.
- Make `USABLE_ALT_DEG` an alias of `site-store.ts`'s
  `DEFAULT_USABLE_ALTITUDE_DEG` instead of a second literal 30, so the pure
  altitude fallback cannot drift from the persisted default.
- Move `ObservingSites.tsx` to `features/settings/`. Its only consumer is
  `PlannerSettings.tsx` and it renders through `SettingsKit`; the four
  observing-sites modules with cross-feature consumers stay in `shared/`.
- Delete the `features/targets/astro/moon-avoidance.ts` re-export shim and point
  its 19 importers at `@/shared/planner/moon-avoidance`.

Non-test import edges: `shared -> features` 4 to 0, `settings -> targets` 3 to 0.
`targets -> settings` stays at 1 module (`planner-sensor.ts` reading camera
equipment over settings IPC), which no longer forms a cycle.
