+++ apps/desktop/src/features/settings/CatalogueSettingsControl.tsx
@@ -13,7 +13,7 @@ import { Toggle } from '@/ui';
 import {
   PLANNER_CATALOGS,
   type CatalogueId,
-} from '@/features/targets/planner-catalog';
+} from '@/shared/planner/planner-catalog';
 import {
   DEFAULT_ENABLED_CATALOGUES,
   loadDefaultCatalogues,
