+++ apps/desktop/src/features/settings/PlannerSettings.tsx
@@ -36,9 +36,9 @@ import {
   setAltitudeThreshold,
   ALTITUDE_THRESHOLD_MIN,
   ALTITUDE_THRESHOLD_MAX,
+  DEFAULT_USABLE_ALTITUDE_DEG,
 } from '@/shared/planner/altitude-settings';
-import { USABLE_ALT_DEG } from '@/features/targets/planner-altitude';
-import { ObservingSites } from '@/shared/observing-sites/ObservingSites';
+import { ObservingSites } from './ObservingSites';
 import {
   useGuidanceParams,
   saveGuidanceParams,
@@ -48,7 +48,7 @@ import {
   WIDTH_MIN,
   WIDTH_MAX,
 } from '@/shared/planner/guidance-settings';
-import { BANDS, type Band } from '@/features/targets/astro/moon-avoidance';
+import { BANDS, type Band } from '@/shared/planner/moon-avoidance';
 
 export function PlannerSettings() {
   const stored = useAltitudeThreshold();
@@ -81,7 +81,9 @@ export function PlannerSettings() {
       <SettingsSection title={m.settings_planner_altitude_title()}>
         <SettingsRow
           label={m.settings_planner_altitude_label()}
-          info={m.settings_planner_altitude_info({ deg: USABLE_ALT_DEG })}
+          info={m.settings_planner_altitude_info({
+            deg: DEFAULT_USABLE_ALTITUDE_DEG,
+          })}
         >
           <input
             type="number"
