+++ apps/desktop/src/features/targets/TargetsTable.test.tsx
@@ -123,7 +123,7 @@ import {
 import { __setObservingStateForTest } from '@/shared/observing-sites/site-store';
 import type { ObserverSite } from '@/shared/observing-sites/observer-site';
 import type { ObservingNight } from './astro/moon-state';
-import { DEFAULT_MOON_AVOIDANCE } from './astro/moon-avoidance';
+import { DEFAULT_MOON_AVOIDANCE } from '@/shared/planner/moon-avoidance';
 
 function item(
   primaryDesignation: string,
