+++ apps/desktop/src/features/targets/planner-derive.test.ts
@@ -17,7 +17,7 @@ import {
   moonExcludedSpans,
 } from './planner-derive';
 import type { ObserverSite } from '@/shared/observing-sites/observer-site';
-import { BANDS, DEFAULT_MOON_AVOIDANCE } from './astro/moon-avoidance';
+import { BANDS, DEFAULT_MOON_AVOIDANCE } from '@/shared/planner/moon-avoidance';
 import { __resetOppositionCacheForTest } from './astro/opposition';
 import { assertDefined } from '@/test/assertDefined';
