+++ apps/desktop/src/features/targets/table-model.ts
@@ -11,7 +11,7 @@
 
 import type { TargetListItem, TargetObjectType } from '@/bindings/index';
 import { objectTypeLabel } from '@/components/TargetSearch/objectType';
-import { catalogueOf, catalogueLabel } from './planner-catalog';
+import { catalogueOf, catalogueLabel } from '@/shared/planner/planner-catalog';
 import { rowAltitudeFor, type RowAltitude } from './planner-altitude';
 import type { ObservingNight } from './astro/moon-state';
 import {
@@ -19,7 +19,7 @@ import {
   type RowMoonGeometry,
   type RowMoonPlanning,
 } from './astro/row-planning';
-import type { MoonAvoidanceParams } from './astro/moon-avoidance';
+import type { MoonAvoidanceParams } from '@/shared/planner/moon-avoidance';
 import { recommendationLabel } from './FilterBadges';
 import { m } from '@/lib/i18n';
 import type { DimensionAccessor } from '@/lib/grouping';
