+++ apps/desktop/src/features/targets/useTargetsTableRows.ts
@@ -21,7 +21,7 @@ import {
   UNKNOWN_ROW_PLANNING,
   type RowMoonGeometry,
 } from './astro/row-planning';
-import type { MoonAvoidanceParams } from './astro/moon-avoidance';
+import type { MoonAvoidanceParams } from '@/shared/planner/moon-avoidance';
 import type { ObserverSite } from '@/shared/observing-sites/observer-site';
 import {
   buildTargetAccessors,
