+++ apps/desktop/src/shared/planner/planner-catalog.ts
